fix: stabilize text rendering on iOS/macOS
Establish Material and DefaultTextStyle boundaries for Cupertino rendering, fix wood grain seed typing, and add text rendering regression tests.
Co-committed-by: Bayanasar Khurd <83147732+bayanasar@users.noreply.github.com>

+++ lib/materials/wood_slab.dart
@@ -51,7 +51,7 @@ class WabWoodSlab extends StatelessWidget {
         child: CustomPaint(
           painter: WabWoodGrain(
             isDark: dark,
-            seed: seed,
+            seed: seed.toDouble(),
             showKnot: showKnot,
           ),
           foregroundPainter: _WoodSlabEdgePainter(dark: dark),
